--- server/domain/orgRepo.go
@@ -1,6 +1,0 @@
-package domain
-
-type OrgRepo interface {
-	GetOrg(OrgId) (Organization, error)
-	GetOrgsForUser(UserId) ([]Organization, error)
-}
