+++ server/domain/user.go
@@ -1,0 +1,8 @@
+package domain
+
+type UserId string
+
+type User struct {
+	Id   UserId
+	Name string
+}
